Improved submission error handling. Fixed player physics. Fixed delta_t multiplication. Converted ratings to doubles.

+++ IPCG/src/Model/GameObject.java
@@ -22,7 +22,7 @@ public abstract class GameObject {
 
 	public abstract Object clone(boolean placeAtEnd);
 
-	public abstract float rate();
+	public abstract double rate();
 
 	public abstract Point getStartPoint();
